% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakr.R
\docType{package}
\name{earthquakr}
\alias{earthquakr}
\alias{earthquakr-package}
\title{earthquakr: A visualization package for world earthquake histories}
\description{
This package maps and timelines earthquakes.
}
