% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Popup Label for Earthquake Map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The \code{\link{quakes}} data frame.}
}
\value{
A vector with the HTML-formatted labels. You should include this
vector with the data frame that is sent to \code{\link{eq_map}}.
}
\description{
\code{eq_create_label} creates a more descriptive and HTML-formatted popup
label to be used in \code{\link{eq_map}}.
}
\details{
This function creates a vector of HTML-formatted labels using supplied
\code{\link{quakes}} data.  The function creates lines of the format
\strong{Label:} \code{ value} from the following variables in the \code{quakes}
data set:
\itemize{
  \item DATE
  \item LOCATION_NAME (as cleaned in the \code{\link{eq_location_clean}}
  function).
  \item EQ_PRIMARY (earthquake magnitude)
  \item TOTAL_DEATHS
}
Any of the above variables with missing/NA values are skipped in the label.
}
\examples{
library(dplyr); library(lubridate)
## Create a popup_text variable and use that for the label
quakes <- eq_load_clean_data() \%>\%
 dplyr::filter(COUNTRY == 'JAPAN') \%>\%
 dplyr::filter(lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.))

eq_map(quakes, annot_col = 'popup_text')
}
