% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Earthquake Location Field}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A data frame of NOAA significant earthquake data, similar to what
can be loaded with \code{data(quakes)}. At a minimum, the data needs to
have \code{LOCATION_NAME} and \code{COUNTRY} features to use this function.}
}
\value{
A \code{tbl_df} with the same supplied data, but with the
  \code{LOCATION_NAME} variable cleaned up to remove the country name that is
  supplied in the \code{COUNTRY} variable (unless the country name is the
  \strong{only} word in the \code{LOCATION_NAME} observation), and convert
  the remainder of the words in \code{LOCATION_NAME} to Title Case.
}
\description{
\code{eq_location_clean} takes a data frame of NOAA earthquake data and cleans
the \code{LOCATION_NAME} field.
}
\details{
This function cleans the \code{LOCATION_NAME} observation in a NOAA
Significant Earthquakes data set.  It removes the country name (of form
\code{COUNTRY:}) from the \code{LOCATION_NAME} data (unless the country is
the only name present in that field), and converts the remainder of the field
to Title Case. This will make the location name easier to read when plotting
and mapping.
}
\examples{
data(quakes)
quakes_loc_clean <- eq_location_clean(quakes)
}
