% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{eq_map}
\alias{eq_map}
\title{Interactive Earthquake Map}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{The \code{\link{quakes}} data frame. For best results, you should
filter to a single country or nearby countries, and preferably filter down
to a relatively narrow date range (maybe several decades at most).}

\item{annot_col}{The column to use for the popup annotation. You may use a
single column; otherwise create a more useful label first by using the
\code{\link{eq_create_label}} function.}
}
\value{
Returns an interactive \code{leaflet} map.
}
\description{
\code{eq_map} creates an interactive \code{\link[leaflet]{leaflet}} map
showing the location of earthquakes in the given \code{quakes} data set.
}
\details{
This function shows an interactive map of the location of the earthquakes in
the given \code{\link{quakes}} data.  The size of the circles are
proportional to the magnitude of the earthquakes (in the \code{EQ_PRIMARY})
variable. The map is interactive, and when you click on a link, the popup
shows the annotation as specified by the \code{annot_col} variable.
}
\examples{
library(lubridate); library(dplyr)

## use 'DATE' as annotation
quakes <- eq_load_clean_data() \%>\%
 dplyr::filter(COUNTRY == 'JAPAN') \%>\%
 dplyr::filter(lubridate::year(DATE) >= 2000)

eq_map(quakes, annot_col = 'DATE')


## Create a popup_text variable and use that for the label
quakes <- eq_load_clean_data() \%>\%
 dplyr::filter(COUNTRY == 'JAPAN') \%>\%
 dplyr::filter(lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.))

eq_map(quakes, annot_col = 'popup_text')
}
\seealso{
\code{\link{eq_create_label}} to create a more useful popup
  annotation in the \code{quakes} data frame.
}
