% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_timeline.R
\name{eq_timeline}
\alias{eq_timeline}
\title{Produce Earthquake Timeline Plots}
\usage{
eq_timeline(df, countries = "USA", date_min = "2000-01-01",
  date_max = "2018-01-01", label_n = 0)
}
\arguments{
\item{df}{The cleaned earthquake data to timeline. See
\code{\link{eq_load_clean_data}} for more information.}

\item{countries}{Default \code{USA}.  May be a single country or a character
vector of countries (countries stacked on y-axis).}

\item{date_min}{Default \code{'2000-01-01'}. Minimum date for timeline in
\code{YYYY-mm-dd} format. May pass in a \code{Date}, a \code{POSIXct}, or
\code{character} vector which can be coerced to a \code{Date}.}

\item{date_max}{Default \code{'2018-01-01'}. Maximum date for timeline in
\code{YYYY-mm-dd} format. May pass in a \code{Date}, a \code{POSIXct}, or
\code{character} vector which can be coerced to a \code{Date}.}

\item{label_n}{Default \code{0}. If \code{label_n == 0}, no labels are
printed on the timeline. Otherwise, the number of labels specified as the
value for this parameter are printed. Suggest keeping this number less
than or equal to 10.}
}
\value{
A \code{ggplot2} object, showing the earthquake timeline.
}
\description{
\code{eq_timeline} is a function to easily create a timeline plot with or
without labels using NOAA Significant Earthquake data.
}
\details{
This function is a wrapper function that makes it easy to create a timeline
plot of the NOAA Significant earthquakes for a given country or countries. It
also allows the user to specify (or not!) the number of earthquakes to label.
This function makes it easy to use the recommended defaults for
\code{\link{geom_timeline}}, \code{\link{geom_timeline_label}}, and
\code{\link{theme_eq}}. The benefit of using this function is that it
saves the user a significant amount of typing and data filtering, though
it gives up control in specifics of how the timelines are displayed.
}
\examples{
# An example with a single country and no labels, showing two
# different classes of date* parameters being passed in.
library(ggplot2); library(dplyr)
quakes <- eq_load_clean_data()
quakes \%>\% eq_timeline(countries = 'USA', date_min = as.Date('1995-01-01'),
                      date_max = as.POSIXct('2015-01-01'), label_n = 0)

# An example with multiple countries and 5 labels per country.
quakes <- eq_load_clean_data()
quakes \%>\% eq_timeline(countries = c('JAPAN', 'INDONESIA', 'HAITI'),
                      date_min = '2000-01-01', date_max = '2017-01-01',
                      label_n = 5)
}
