% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Timeline of Earthquakes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{other arguments passed on to \code{\link{layer}}.}

\item{x}{Required, recommend \code{DATE}}

\item{y}{Optional, recommend \code{COUNTRY}}

\item{size}{Optional, recommend \code{EQ_PRIMARY} (magnitude)}

\item{color}{Optional, recommend \code{TOTAL_DEATHS}}
}
\description{
\code{geom_timeline} shows a timeline of NOAA Significant earthquakes,
plotting individual countries along the y-axis and dates along the x-axis.
The size of the points is relatative to the earthquakes' magnitude, and
the color is related to the total number of deaths.
}
\details{
It is highly recommended that this geom is not used in isolation. For
best results, use it with the \code{\link{eq_timeline}} wrapper function.
}
\examples{
library(dplyr)
quakes <- eq_load_clean_data()

quakes \%>\% dplyr::filter(COUNTRY \%in\% c('USA', 'UK')) \%>\%
  dplyr::filter(DATE > '2000-01-01') \%>\%
  ggplot() +
  geom_timeline(aes(x = DATE, y = COUNTRY, color = TOTAL_DEATHS,
                    size = EQ_PRIMARY)) +
  ggplot2::scale_size_continuous(name = 'Richter scale value') +
  ggplot2::scale_color_continuous(name = '# of Deaths') +
  theme_eq()
}
