% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quakes}
\alias{quakes}
\title{NOAA Significant Earthquakes}
\format{A data frame of 5934 observations of 47 variables.  For variable
  definitions please see the "Event Variable Definitions" link at the url
  below. These data were pulled on April 1, 2017; see source for any updated
  data after this date.}
\source{
National Geophysical Data Center / World Data Service (NGDC/WDS):
  Significant Earthquake Database. National Geophysical Data Center, NOAA.
  doi:10.7289/V5TD9V7K
  \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
quakes
}
\description{
"The Significant Earthquake Database contains information on destructive
earthquakes from 2150 B.C. to the present that meet at least one of the
following criteria: Moderate damage (approximately $1 million or more), 10 or
more deaths, Magnitude 7.5 or greater, Modified Mercalli Intensity X or
greater, or the earthquake generated a tsunami."
}
\keyword{datasets}
