% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eq.R
\name{theme_eq}
\alias{theme_eq}
\title{Earthquake Timeline Theme}
\usage{
theme_eq(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_size}{Default \code{11}, base size of text}

\item{base_family}{Default \code{sans}, text family}
}
\description{
Makes your earthquakr timeline plots look pretty.
}
\examples{
library(dplyr); library(ggplot2)
quakes <- eq_load_clean_data()

quakes \%>\% dplyr::filter(COUNTRY \%in\% c('USA', 'UK')) \%>\%
  dplyr::filter(DATE > '2000-01-01') \%>\%
  ggplot2::ggplot() +
  geom_timeline(aes(x = DATE, y = COUNTRY, color = TOTAL_DEATHS,
                    size = EQ_PRIMARY)) +
  ggplot2::scale_size_continuous(name = 'Richter scale value') +
  ggplot2::scale_color_continuous(name = '# of Deaths') +
  theme_eq()
}
